# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/publication": {
            "post": {
                "description": "Post publication",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "multipart/form-data"
                ],
                "tags": [
                    "Publication"
                ],
                "summary": "Post publication",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "file"
                        },
                        "collectionFormat": "csv",
                        "description": "files",
                        "name": "files",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "text",
                        "in": "formData"
                    }
                ],
                "responses": {}
            }
        },
        "/publication/{id}": {
            "get": {
                "description": "Get publication",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Publication"
                ],
                "summary": "Get publication",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "publication id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "delete": {
                "description": "Delete publication",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "multipart/form-data"
                ],
                "tags": [
                    "Publication"
                ],
                "summary": "Delete publication",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    }
}